StartWithChat/requirements.txt
langchain_community
langchain
langchain_huggingface
python-dotenv
faiss-cpu
sentence-transformers
pypdf
